moving dde setting file to another location
